Set QOS inside consumer loop
When we need to reconnect, it would reconnect the consumer without the
QOS being set. This is because the QOS is not being set inside the main
loop. So when that loop returns to the beginning, we get a consumer with
no QOS being set. This moves the QOS to being set inside the main loop
so if we do return to the beginning of the loop, we also set the QOS
again as well.