Add gradually increasing connection delay
This adds a gradually increasing connection backoff delay so that when
the application fails to connect many times consecutively it will
continually retry with an increasing wait time between succeessive
tries. This adds the ability for the application to immediately try to
reconnect when it gets disconnected and to be smart about not
immediately trying to reconnect if the last 10 times failed as well.