Update and add handled exceptions
This adds some exceptions as well as updates some exceptions to be a
more high-level exception as a lot of exceptions for slightly different
things are very similar. So instead of catching multiple very similar
exceptions, we can just catch a single exception and not have to worry
about knowing every single possible exception that can be thrown.