Add more debug logging
When the consumer starts, it initially was sitting on a message for
"Establishing connection..." and it seemingly never got past the
connection creation from what the logs were outputting even though it
did get past it just fine. This adds more logging to make it apparent
that it gets past the connection creation just fine and is no longer
printing any messages to the log because it's "Waiting for messages...".