Relocate queue declare for consumers
This relocates the queue declare when using a consumer. This makes it so
that when the consumer attempts to reconnect, if the AMQP server got
wiped the queue that was previously declared will not be available
anymore. So this guarantees the queue that is supposed to be listened to
is going to be there when it connects.